+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 - Fix problema en menú de indicadores (se ocultan ítems en ciertas resoluciones)
 - Fix a performance de publicación (se demora al 88% en ciertas cartografías).
 - La tabla de boundaries permite indicar cuáles deben ser destacados en el visor en la parte superior izquierda.
+- Ofrece guías al no haber cartografías.
+- Fix a capas con deckgl.
+- Escala en mapas con leaflet.
+- Links a ayuda en PDF en las burbujas de ayuda.
+
 
 v4.3 (2025-04-14)
 - Mejoras en manejo de error al publicar.